pyproject.toml
[tool.poetry]
name = "ton-dex-tegrofinance-web-backend"
version = "0.1.0"
description = "Backend for TON-DEX-tegrofinance-web."
authors = ["Dmitrii Kopeikin <kopeikindp@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ton_dex_tegrofinance_web_backend"}]

[tool.poetry.dependencies]
python = "^3.10,<3.11"
httpx = "^0.25.2"
pydantic-settings = "^2.1.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
tonsdk = "^1.0.13"
aiohttp = "^3.9.1"
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
aiosqlite = "^0.19.0"
sqlalchemy-utils = "^0.41.1"
pyjwt = "^2.8.0"
apscheduler = "^3.10.4"
pytonapi = "^0.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
